Move some data structures from `sc_network::config` and `sc_network::request_responses` into new `sc-network-common` crate
