+++ Cargo.toml
@@ -42,6 +42,7 @@ members = [
 	"client/informant",
 	"client/keystore",
 	"client/network",
+	"client/network/common",
 	"client/network-gossip",
 	"client/network/sync",
 	"client/network/test",
