+++ client/network/common/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+description = "Substrate network common"
+name = "sc-network-common"
+version = "0.10.0-dev"
+license = "GPL-3.0-or-later WITH Classpath-exception-2.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+edition = "2021"
+homepage = "https://substrate.io"
+repository = "https://github.com/paritytech/substrate/"
+documentation = "https://docs.rs/sc-network-sync"
+readme = "README.md"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[build-dependencies]
+prost-build = "0.9"
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "3.0.0", features = [
+    "derive",
+] }
+futures = "0.3.21"
+libp2p = "0.44.0"
+sc-peerset = { version = "4.0.0-dev", path = "../../peerset" }
+smallvec = "1.8.0"
